+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 2.2.1, 10/1/2021
+
+### Added
+
+Callback function can implement ServiceExceptionHandler to catch exception. It adds the onError() method.
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 2.1.1, 9/10/2021
 
 ### Added
@@ -14,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 2. Standardized serializers for default case, snake_case and camelCase
 3. Support of EventEnvelope as input parameter in TypedLambdaFunction so application function can inspect event's metadata
 4. Application can subscribe to life cycle events of other application instances
+5. ServiceExceptionHandler for callback function to catch exception
 
 ### Removed